--- a/ext/PyPlotExt.jl
+++ b/ext/PyPlotExt.jl
@@ -33,7 +33,7 @@ function plotit(x, y, waterflows_output::Tuple, dem)
 end
 
 function plotdir(x, y, dir)
-    vecfield = WWF.dir2vec.(dir)
+    vecfield = WWF.dir2vec.(dir, true)
     vecfieldx = [v[1] for v in vecfield]
     vecfieldy = [v[2] for v in vecfield]
     quiver(repeat(x,1, length(y)), repeat(y,length(x),1), vecfieldx, vecfieldy)
@@ -78,8 +78,8 @@ function plotbnds(x,y,bnds)
 end
 
 function plotlakedepth(x, y, dem)
-    area, slen, dir, nout, nin, pits  = waterflows(dem)
-    demf = WWF.fill_dem(dem, pits, dir)
+    area, slen, dir, nout, nin, sinks, pits  = waterflows(dem)
+    demf = WWF.fill_dem(dem, sinks, dir)
     heatmap(x, y, demf.-dem)
 end