--- a/src/postproc.jl
+++ b/src/postproc.jl
@@ -132,11 +132,14 @@ end
 Fill the pits of a DEM (apply this after applying "drainpits",
 which is done by default in `waterflows`). Returns the filled DEM.
 
-Note, this is *not* needed as pre-processing step to use the flow-routing
-function `waterflows`.
-
-This uses a tree traversal to fill the DEM. It does it depth-first (as it
-is easier) which may lead to a stack overflow on a large DEM.
+Notes:
+- this is *not* needed as pre-processing step to use the flow-routing
+  function `waterflows`.
+- routing on a filled DEM will not produce exactly the same flow pattern: on
+  the shores of lakes streams which entered the lake can now go the other way.
+  I suspect on most DEMs the differences will be very minimal.
+- This uses a tree traversal to fill the DEM. It does it depth-first (as it
+  is easier) which may lead to a stack overflow on a large DEM.
 """
 function fill_dem(dem, sinks, dir; small=0)
     dem = copy(dem)
@@ -150,9 +153,9 @@ end
 # which has lower elevation than the previous one, it will set that point's elevation
 # and all upstream points' elevation the elevation of the first point.
 function _fill_ij!(ele, dem, ij, dir, small)
-    if ele > dem[ij]
+    if ele >= dem[ij]
+        ele += 2*eps(ele)
         dem[ij] = ele
-        ele += small
     else
         ele = dem[ij]
     end