--- a/test/postproc.jl
+++ b/test/postproc.jl
@@ -5,7 +5,7 @@
     ys = -0.5:dx:3.0
     dem = dem1.(xs, ys', withpit=true)
     area, slen, dir, nout, nin, sinks, pits, c, bnds = WWF.waterflows(dem, bnd_as_sink=true)
-    demf = WWF.fill_dem(dem, pits, dir)
+    demf = WWF.fill_dem(dem, sinks, dir)
     @test sum(demf.-dem) ≈ 2.1499674517313414
     @test sum(demf.-dem .> 0) == 5
     @test all([c[pits[cc]]==cc  for cc=axes(pits)[1]]) # pit in catchment of same color